extract common spring boot beans to this.
