--- src/main/java/io/nayasis/basica/spring/web/client/function/ResponseChecker.java
@@ -1,6 +1,0 @@
-package io.nayasis.basica.spring.web.client.function;
-
-@FunctionalInterface
-public interface ResponseChecker<T> {
-    void check( T response );
-}
