+++ src/main/java/io/nayasis/spring/extension/config/message/CustomMessageSource.java
@@ -1,0 +1,32 @@
+package io.nayasis.spring.extension.config.message;
+
+import io.nayasis.basica.model.Messages;
+import io.nayasis.basica.validation.Validator;
+import org.springframework.context.MessageSource;
+import org.springframework.context.MessageSourceResolvable;
+import org.springframework.context.NoSuchMessageException;
+
+import java.util.Locale;
+
+public class CustomMessageSource implements MessageSource {
+
+    private final Object[] EMPTY_ARGS = {};
+
+    @Override
+    public String getMessage( String code, Object[] args, String defaultMessage, Locale locale ) {
+        return Messages.get( locale, code, Validator.nvl( args, EMPTY_ARGS ) );
+    }
+
+    @Override
+    public String getMessage( String code, Object[] args, Locale locale ) throws NoSuchMessageException {
+        return getMessage( code, args, null, locale );
+    }
+
+    @Override
+    public String getMessage( MessageSourceResolvable resolvable, Locale locale ) throws NoSuchMessageException {
+        String[] codes = resolvable.getCodes();
+        Object[] args = Validator.nvl( resolvable.getArguments(), EMPTY_ARGS );
+        return Messages.get( locale, codes[0], args );
+    }
+
+}
