+++ src/main/java/io/nayasis/spring/extension/web/client/function/ResponseChecker.java
@@ -1,0 +1,6 @@
+package io.nayasis.spring.extension.web.client.function;
+
+@FunctionalInterface
+public interface ResponseChecker<T> {
+    void check( T response );
+}
